Add files via upload
This is a program that administers and grades a math quiz with randomly generated questions of several different kinds. I targeted the quiz at an elementary school student who is learning basic arithmetic. The questions are asking the student to perform simple addition, subtraction, multiplication, and division problems.

For each question, the type of problem is selected at random, and then the numbers that occur in the problem is randomly generated.